packets now use int for celestialType and silverfishes use various variants instead of being only EntitySilverfish

+++ src/main/java/vbonedra/hostiles_are_too_easy/network/S2CResponseCelestialType.java
@@ -8,7 +8,7 @@ import vbonedra.hostiles_are_too_easy.util.CelestialTypeCache;
 public class S2CResponseCelestialType implements Packet {
     private final int entityId;
     private final int celestialType;
-
+    // TODO?: maybe split into 2 packets, one that sends byte and another that sends int, this might improve performance if int for all types is bad
     public S2CResponseCelestialType(int entityId, int celestialType) {
         this.entityId = entityId;
         this.celestialType = celestialType;
@@ -16,13 +16,13 @@ public class S2CResponseCelestialType implements Packet {
 
     public S2CResponseCelestialType(PacketByteBuf buf) {
         this.entityId = buf.readInt();
-        this.celestialType = buf.readByte();
+        this.celestialType = buf.readInt();
     }
 
     @Override
     public void write(PacketByteBuf buf) {
         buf.writeInt(this.entityId);
-        buf.writeByte(this.celestialType);
+        buf.writeInt(this.celestialType);
     }
 
     @Override
